Refactor the MsSamr mixin to split it out
